==> notebooks/01_loading and EDA.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df=pd.read_csv("../datasets/Telco_Customer_Churn_Dataset.csv")

for col in df:
    df.rename(columns={col:col.lower()},inplace=True)

df.info()

df.nunique()

for col in df.columns:
    print(col,":",df[col].unique())

for col in df.columns:
    print(col)

customerID is not required for our analysis drop it

df=df.drop("customerid",axis=1)

df1=df.copy(deep=True)

df1.info()

No internet service and No phone service is simlar to no. So convert it into No

temp_list=['onlinesecurity','onlinebackup','deviceprotection','techsupport','streamingtv','streamingmovies']
for temp_col in temp_list:
    df1[temp_col]=df1[temp_col].replace('No internet service','No')

for col in df1.columns:
    print(col,":",df1[col].unique())

df1.info()

We can observe that total charges is in object format convert that to float

df.totalcharges.isnull().sum()

df1["totalcharges"]=pd.to_numeric(df1["totalcharges"])

df1["totalcharges"]=pd.to_numeric(df1["totalcharges"],errors='coerce')


df1.dropna(inplace=True)
df1.isnull().sum()

Categorical variable encoding

for col in df1.columns:
    print(col,":",df1[col].unique())

df1['gender'] = df1['gender'].map({'Male': 1, 'Female': 0})

for col in df1.columns:
    if df1[col].nunique()<3 and col not in ['gender','seniorcitizen']:#because gender and seniorcitizen is already encoded
        df1[col] = df1[col].map({'Yes': 1, 'No': 0})

# multiplelines : ['No phone service' 'No' 'Yes']
df1['multiplelines'] = df1['multiplelines'].map({'Yes': 1, 'No': 0,'No phone service':2})

# internetservice : ['DSL' 'Fiber optic' 'No']
df1['internetservice'] = df1['internetservice'].map({'DSL': 1, 'Fiber optic': 2,'No':0})

# contract : ['Month-to-month' 'One year' 'Two year']
df1['contract'] = df1['contract'].map({'Month-to-month': 0, 'One year': 1,'Two year':2})

# paymentmethod : ['Electronic check' 'Mailed check' 'Bank transfer (automatic)' 'Credit card (automatic)']
df1['paymentmethod'] = df1['paymentmethod'].map({'Electronic check': 0, 'Mailed check': 1,'Bank transfer (automatic)':2,'Credit card (automatic)':3})

for col in df1.columns:
    print(col,":",df1[col].unique())

Outlier detection

def find_outliers_iqr(data, column):
    Q1 = data[column].quantile(0.25)
    Q3 = data[column].quantile(0.75)
    IQR = Q3 - Q1
    lower = Q1 - 1.5 * IQR
    upper = Q3 + 1.5 * IQR
    outliers = data[(data[column] < lower) | (data[column] > upper)]
    return outliers[[column]]

# Example usage:
outliers_monthly = find_outliers_iqr(df1, 'monthlycharges')
outliers_total = find_outliers_iqr(df1, 'totalcharges')
outliers_tenure = find_outliers_iqr(df1, 'tenure')


outliers_monthly

outliers_total

outliers_tenure

Conclusion: NO outliers detected

df1.to_csv('cleaned.csv', index=False)
