==> notebooks/02_test_train_split.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split, StratifiedKFold

df=pd.read_csv("../datasets/cleaned.csv")

x,y=df.drop('churn',axis=1),df.churn
x_train, x_test, y_train, y_test = train_test_split(
    x, y,
    test_size=0.2,       
    random_state=42,     
    stratify=y           # preserves class balance
)

#to balance the remaning dataset
from imblearn.over_sampling import SMOTE
smote=SMOTE(sampling_strategy='minority')
x_train_bal,y_train_bal=smote.fit_resample(x_train,y_train)

x_train_bal.shape

x_train_bal.columns

x_train_bal.to_csv("../datasets/x_train_bal.csv",index=False)
y_train_bal.to_csv("../datasets/y_train_bal.csv",index=False)
x_test.to_csv("../datasets/x_test.csv",index=False)
y_test.to_csv("../datasets/y_test.csv",index=False)
