==> notebooks/05_model_training.ipynb <==
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, GridSearchCV, RandomizedSearchCV,cross_val_score
from sklearn.metrics import accuracy_score, classification_report, roc_auc_score

selected_features=['partner',
 'onlinebackup',
 'tenure',
 'contract',
 'monthlycharges',
 'totalcharges',
 'techsupport', 'internetservice',
 'multiplelines']


x_train_bal=pd.read_csv("../datasets/x_train_bal.csv")
y_train_bal=pd.read_csv("../datasets/y_train_bal.csv")
x_train_bal=x_train_bal[selected_features]

x_train_bal.head()

y_train_bal.value_counts()

Using GridSearch to find best parameters.

param_grid = {
    'n_estimators': [100, 200, 300],
    'max_depth': [10, 20, 30, None],
    'min_samples_split': [2, 5, 10],
    'min_samples_leaf': [1, 2, 4],
    'max_features': ['sqrt', 'log2'],
    'bootstrap': [True, False],
    'criterion': ['gini', 'entropy'],
    'class_weight': ['balanced', None],
    'max_leaf_nodes': [None, 20, 50],
    'min_impurity_decrease': [0.0, 0.01]
}

grid_skf = GridSearchCV(
    estimator=RandomForestClassifier(random_state=42),
    param_grid=param_grid,
    cv=3,               # <-- use StratifiedKFold here
    scoring='accuracy',
    verbose=3,
    n_jobs=-1 
)

grid_skf.fit(x_train_bal, y_train_bal)
print("Best params (StratifiedKFold):", grid_skf.best_params_)

 {'bootstrap': True, 'class_weight': 'balanced', 'criterion': 'gini', 'max_depth': None, 'max_features': 'sqrt', 'max_leaf_nodes': None, 'min_impurity_decrease': 0.0, 'min_samples_leaf': 2, 'min_samples_split': 2, 'n_estimators': 100, 'verbose': 0, 'warm_start': False}

grid_skf.best_params_

grid_skf2 = GridSearchCV(
    estimator=RandomForestClassifier(random_state=42),
    param_grid=param_grid,
    cv=5,               # <-- use StratifiedKFold here
    scoring='accuracy',
    verbose=3,
    n_jobs=-1 
)
grid_skf2.fit(x_train_bal, y_train_bal)
print("Best params (StratifiedKFold):", grid_skf.best_params_)

grid_skf2.best_score_

grid_skf2.best_params_

rf_model = RandomForestClassifier(
    bootstrap=True,
    class_weight='balanced',
    criterion='gini',
    max_depth=20,
    max_features='sqrt',
    max_leaf_nodes=None,
    min_impurity_decrease=0.0,
    min_samples_leaf=2,
    min_samples_split=2,
    n_estimators=200,
    verbose=0,
    warm_start=False
)

rf_model.fit(x_train_bal, y_train_bal)

x_test=pd.read_csv("../datasets/x_test.csv")
y_test=pd.read_csv("../datasets/y_test.csv")
x_test=x_test[selected_features]

y_pred_rf = rf_model.predict(x_test)
y_prob_rf = rf_model.predict_proba(x_test)[:, 1]

test_accuracy = accuracy_score(y_test, y_pred_rf)
test_auc = roc_auc_score(y_test, y_prob_rf)

print(f"\nTest Accuracy: {round(test_accuracy, 4)}")
print(f"Test AUC: {round(test_auc, 4)}")

from sklearn.model_selection import RandomizedSearchCV
random_search = RandomizedSearchCV(
    estimator=RandomForestClassifier(random_state=42),
    param_distributions=param_grid,
    n_iter=200,                # number of parameter settings sampled
    cv=5,                     # 5-fold cross-validation
    verbose=2,
    random_state=42,
    n_jobs=-1                 # use all processors
)
random_search.fit(x_train_bal, y_train_bal)

random_search.best_score_

random_search.best_params_

Since GridSearchCV was taking very long time execute.
Using RandomizedSearchCV for model tunning

random_search2 = RandomizedSearchCV(
    estimator=RandomForestClassifier(random_state=42),
    param_distributions=param_grid,
    n_iter=2000,                # number of parameter settings sampled
    cv=5,                     # 5-fold cross-validation
    verbose=2,
    random_state=42,
    n_jobs=-1                 # use all processors
)
random_search2.fit(x_train_bal, y_train_bal)

random_search2.best_score_

random_search2.best_params_

selected_parameters={
 'n_estimators': 200,        
 'min_samples_split': 2,              
 'min_samples_leaf': 2,               
 'min_impurity_decrease': 0.0,        
 'max_leaf_nodes': None,               
 'max_features': 'log2',              
 'max_depth': 20,        
 'criterion': 'gini',                 
 'class_weight': None,                
 'bootstrap': True                     
}


final_model=RandomForestClassifier(**selected_parameters)

final_fit=final_model.fit(x_train_bal,y_train_bal)

x_test=pd.read_csv("../datasets/x_test.csv")
y_test=pd.read_csv("../datasets/y_test.csv")
x_test=x_test[selected_features]

y_predict=final_fit.predict(x_test)

y_proba = final_fit.predict_proba(x_test)[:, 1]  # probability of class 1

# Accuracy
test_accuracy = accuracy_score(y_test, y_predict)

# ROC AUC (using probabilities)
test_accuracy = accuracy_score(y_test, y_predict)
test_auc = roc_auc_score(y_test, y_proba)

print(f"\nTest Accuracy: {round(test_accuracy, 4)}")
print(f"Test AUC: {round(test_auc, 4)}")

To save the model

import joblib
joblib.dump(final_fit, '../churn_predictor.pkl')
