==> notebooks/06_model_evaluation.ipynb <==
MODEL EVALUATION

import pandas as pd
import joblib
from sklearn.metrics import accuracy_score, classification_report, roc_auc_score,confusion_matrix

selected_features=['partner',
 'onlinebackup',
 'tenure',
 'contract',
 'monthlycharges',
 'totalcharges',
 'techsupport', 'internetservice',
 'multiplelines']


x_test=pd.read_csv("../datasets/x_test.csv")
y_test=pd.read_csv("../datasets/y_test.csv")
x_test=x_test[selected_features]

model = joblib.load('../churn_predictor.pkl')

y_predicted=model.predict(x_test)

report=classification_report(y_test,y_predicted)

print("Model evaluation:")
print(report)

cm = confusion_matrix(y_test, y_predicted)

print("Confusion Matrix:")
print("True Positive  :",cm[0][0])
print("True Negative  :",cm[1][1])
print("False Positive :",cm[0][1])
print("False Negative :",cm[1][0])
